+++ CHANGES.rst
@@ -8,9 +8,20 @@ Changes in current version of Apache Libcloud
 Compute
 ~~~~~~~
 
+- Added new CloudScale.ch driver
+
+- [dimensiondata] Added support for 2.4 API, added support for image import, cloning. Add feature for changing NIC VLANs, add feature for changing
+  NIC order for a server.
+  (GITHUB-953)
+  [Samuel Chong]
+
 - [ec2] Add US-EAST2 (Ohio)
   (GITHUB-946)
   [Matthew Harris]
+
+- [google compute] Fix to allow multiple node creation with subnets
+  (GITHUB-949)
+  [John Baublitz]
 
 Changes in Apache Libcloud 1.4.0
 --------------------------------